Creation of Channels
The bot will now search for the existence of specific channels when being created. If they do not exist, it will create an appropriate group, text channel, and voice channel to perform actions in. Will likely need to move this to a listener, in the event that something is deleted that the bot expects to exist.